move sorting into view container functionality
